rename `testdata` to `fixtures`
