+++ .gitignore
@@ -3,5 +3,5 @@ Cargo.lock
 _test
 
 # May be created during tests, but should not be present
-tests/testdata/sparse_registry_cache/cargo_home/registry/index/index.crates.io-6f17d22bba15001f/.cache/cr/at/crates-index
-/tests/testdata/git-registry/
+tests/fixtures/sparse_registry_cache/cargo_home/registry/index/index.crates.io-6f17d22bba15001f/.cache/cr/at/crates-index
+/tests/fixtures/git-registry/
