+++ src/bare_index.rs
@@ -538,7 +538,7 @@ mod test {
     }
 
     fn shared_index() -> Index {
-        let index_path = "tests/testdata/git-registry";
+        let index_path = "tests/fixtures/git-registry";
         if is_ci::cached() {
             Index::new_cargo_default()
                 .expect("CI has just cloned this index and its ours and valid")
