+++ tests/fixtures/autocfg.txt
