+++ tests/fixtures/crates-index.txt
